Get rid of the old non-shutdown versions of the fuzzing targets. The shutdown versions were a superset, so the old are not necessary anymore.
