+++ fuzz/Cargo.toml
@@ -32,15 +32,3 @@ name = "map"
 path = "fuzz_targets/map.rs"
 test = false
 doc = false
-
-[[bin]]
-name = "queue-with-shutoff"
-path = "fuzz_targets/queue_with_shutoff.rs"
-test = false
-doc = false
-
-[[bin]]
-name = "map-with-shutoff"
-path = "fuzz_targets/map_with_shutoff.rs"
-test = false
-doc = false
