+++ fuzz/fuzz_targets/map.rs
@@ -2,37 +2,40 @@
 
 use libfuzzer_sys::arbitrary::Arbitrary;
 use libfuzzer_sys::fuzz_target;
+use rand::SeedableRng;
 use sequential_storage::{
     map::{MapError, StorageItem},
-    mock_flash::MockFlashBase,
+    mock_flash::{MockFlashBase, MockFlashError, WriteCountCheck},
 };
 use std::{collections::HashMap, ops::Range};
 
 fuzz_target!(|data: Input| fuzz(data));
 
-#[derive(Arbitrary, Debug)]
+#[derive(Arbitrary, Debug, Clone)]
 struct Input {
+    seed: u64,
+    fuel: u16,
     ops: Vec<Op>,
 }
 
-#[derive(Arbitrary, Debug)]
+#[derive(Arbitrary, Debug, Clone)]
 enum Op {
     Store(StoreOp),
     Fetch(u8),
 }
 
-#[derive(Arbitrary, Debug)]
+#[derive(Arbitrary, Debug, Clone)]
 struct StoreOp {
     key: u8,
     value_len: u8,
 }
 
 impl StoreOp {
-    fn into_test_item(self) -> TestItem {
+    fn into_test_item(self, rng: &mut impl rand::Rng) -> TestItem {
         TestItem {
             key: self.key,
-            value: (0..self.value_len as usize)
-                .map(|_| rand::random::<u8>())
+            value: (0..(self.value_len % 8) as usize)
+                .map(|_| rng.gen())
                 .collect(),
         }
     }
@@ -87,55 +90,134 @@ fn fuzz(ops: Input) {
     const WORD_SIZE: usize = 4;
     const WORDS_PER_PAGE: usize = 256;
 
-    let mut flash = MockFlashBase::<PAGES, WORD_SIZE, WORDS_PER_PAGE>::default();
+    let mut flash = MockFlashBase::<PAGES, WORD_SIZE, WORDS_PER_PAGE>::new(
+        WriteCountCheck::OnceOnly,
+        Some(ops.fuel as u32),
+    );
     const FLASH_RANGE: Range<u32> = 0x000..0x1000;
 
     let mut map = HashMap::new();
     let mut buf = [0; 260]; // Max length of test item serialized, rounded up to align to flash word.
 
-    for op in ops.ops.into_iter() {
-        // println!(
-        //     "==================================================== op: {:?}",
-        //     op,
-        // );
-        match op {
-            Op::Store(op) => {
-                let item = op.into_test_item();
-                match sequential_storage::map::store_item(
-                    &mut flash,
-                    FLASH_RANGE,
-                    &mut buf,
-                    item.clone(),
-                ) {
-                    Ok(_) => {
-                        map.insert(item.key, item.value);
-                    }
-                    Err(MapError::FullStorage) => {}
-                    Err(e) => panic!("{e:?}"),
-                }
-            }
-            Op::Fetch(key) => {
-                let fetch_result = sequential_storage::map::fetch_item::<TestItem, _>(
-                    &mut flash,
-                    FLASH_RANGE,
-                    &mut buf,
-                    key,
-                )
-                .unwrap();
+    let mut rng = rand_pcg::Pcg32::seed_from_u64(ops.seed);
 
-                if let Some(existing_value) = map.get(&key) {
-                    assert_eq!(
-                        fetch_result.as_ref().map(|item| &item.key),
-                        Some(&key),
-                        "Mismatching keys"
-                    );
-                    assert_eq!(
-                        fetch_result.as_ref().map(|item| &item.value),
-                        Some(existing_value),
-                        "Mismatching values"
-                    );
-                } else {
-                    assert_eq!(fetch_result, None)
+    #[cfg(fuzzing_repro)]
+    eprintln!("\n=== START ===\n");
+
+    for op in ops.ops.into_iter() {
+        let mut retry = true;
+        let mut corruption_repaired = false;
+
+        while std::mem::replace(&mut retry, false) {
+            #[cfg(fuzzing_repro)]
+            eprintln!("{}", flash.print_items());
+            #[cfg(fuzzing_repro)]
+            eprintln!("=== OP: {op:?} ===");
+
+            match op.clone() {
+                Op::Store(op) => {
+                    let item = op.into_test_item(&mut rng);
+                    match sequential_storage::map::store_item(
+                        &mut flash,
+                        FLASH_RANGE,
+                        &mut buf,
+                        item.clone(),
+                    ) {
+                        Ok(_) => {
+                            map.insert(item.key, item.value);
+                        }
+                        Err(MapError::FullStorage) => {}
+                        Err(MapError::Storage {
+                            value: MockFlashError::EarlyShutoff(_),
+                            backtrace: _backtrace,
+                        }) => {
+                            match sequential_storage::map::fetch_item::<TestItem, _>(
+                                &mut flash,
+                                FLASH_RANGE,
+                                &mut buf,
+                                item.key,
+                            ) {
+                                Ok(Some(check_item))
+                                    if check_item.key == item.key
+                                        && check_item.value == item.value =>
+                                {
+                                    #[cfg(fuzzing_repro)]
+                                    eprintln!("Early shutoff when storing {item:?}! (but it still stored fully). Originated from:\n{_backtrace:#}");
+                                    // Even though we got a shutoff, it still managed to store well
+                                    map.insert(item.key, item.value);
+                                }
+                                _ => {
+                                    // Could not fetch the item we stored...
+                                    #[cfg(fuzzing_repro)]
+                                    eprintln!("Early shutoff when storing {item:?}! Originated from:\n{_backtrace:#}");
+                                }
+                            }
+                        }
+                        Err(MapError::Corrupted {
+                            backtrace: _backtrace,
+                        }) if !corruption_repaired => {
+                            #[cfg(fuzzing_repro)]
+                            eprintln!(
+                                "### Encountered curruption while storing! Repairing now. Originated from:\n{_backtrace:#}"
+                            );
+
+                            sequential_storage::map::try_repair::<TestItem, _>(
+                                &mut flash,
+                                FLASH_RANGE,
+                                &mut buf,
+                            )
+                            .unwrap();
+                            corruption_repaired = true;
+                            retry = true;
+                        }
+                        Err(e) => panic!("{e:?}"),
+                    }
+                }
+                Op::Fetch(key) => {
+                    match sequential_storage::map::fetch_item::<TestItem, _>(
+                        &mut flash,
+                        FLASH_RANGE,
+                        &mut buf,
+                        key,
+                    ) {
+                        Ok(Some(fetch_result)) => {
+                            let map_value = map
+                                .get(&key)
+                                .expect(&format!("Map doesn't contain: {fetch_result:?}"));
+                            assert_eq!(key, fetch_result.key, "Mismatching keys");
+                            assert_eq!(map_value, &fetch_result.value, "Mismatching values");
+                        }
+                        Ok(None) => {
+                            assert_eq!(None, map.get(&key));
+                        }
+                        Err(MapError::Storage {
+                            value: MockFlashError::EarlyShutoff(_),
+                            backtrace: _backtrace,
+                        }) => {
+                            #[cfg(fuzzing_repro)]
+                            eprintln!(
+                                "Early shutoff when fetching! Originated from:\n{_backtrace:#}"
+                            );
+                        }
+                        Err(MapError::Corrupted {
+                            backtrace: _backtrace,
+                        }) if !corruption_repaired => {
+                            #[cfg(fuzzing_repro)]
+                            eprintln!(
+                                "### Encountered curruption while fetching! Repairing now. Originated from:\n{_backtrace:#}"
+                            );
+
+                            sequential_storage::map::try_repair::<TestItem, _>(
+                                &mut flash,
+                                FLASH_RANGE,
+                                &mut buf,
+                            )
+                            .unwrap();
+                            corruption_repaired = true;
+                            retry = true;
+                        }
+                        Err(e) => panic!("{e:?}"),
+                    }
                 }
             }
         }
